requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.1
django-allauth==0.37.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
wincertstore==0.2
